update: use neo3 codes to replace old
