+++ packages/neon-core/src/sc/InteropService.ts
