+++ packages/neon-core/src/sc/OpCode.ts
@@ -2,6 +2,79 @@ export enum OpCode {
   PUSH0 = 0,
   PUSHF = 0,
   PUSHBYTES1 = 1,
+  PUSHBYTES2 = 2,
+  PUSHBYTES3 = 3,
+  PUSHBYTES4 = 4,
+  PUSHBYTES5 = 5,
+  PUSHBYTES6 = 6,
+  PUSHBYTES7 = 7,
+  PUSHBYTES8 = 8,
+  PUSHBYTES9 = 9,
+  PUSHBYTES10 = 10,
+  PUSHBYTES11 = 11,
+  PUSHBYTES12 = 12,
+  PUSHBYTES13 = 13,
+  PUSHBYTES14 = 14,
+  PUSHBYTES15 = 15,
+  PUSHBYTES16 = 16,
+  PUSHBYTES17 = 17,
+  PUSHBYTES18 = 18,
+  PUSHBYTES19 = 19,
+  PUSHBYTES20 = 20,
+  PUSHBYTES21 = 21,
+  PUSHBYTES22 = 22,
+  PUSHBYTES23 = 23,
+  PUSHBYTES24 = 24,
+  PUSHBYTES25 = 25,
+  PUSHBYTES26 = 26,
+  PUSHBYTES27 = 27,
+  PUSHBYTES28 = 28,
+  PUSHBYTES29 = 29,
+  PUSHBYTES30 = 30,
+  PUSHBYTES31 = 31,
+  PUSHBYTES32 = 32,
+  PUSHBYTES33 = 33,
+  PUSHBYTES34 = 34,
+  PUSHBYTES35 = 35,
+  PUSHBYTES36 = 36,
+  PUSHBYTES37 = 37,
+  PUSHBYTES38 = 38,
+  PUSHBYTES39 = 39,
+  PUSHBYTES40 = 40,
+  PUSHBYTES41 = 41,
+  PUSHBYTES42 = 42,
+  PUSHBYTES43 = 43,
+  PUSHBYTES44 = 44,
+  PUSHBYTES45 = 45,
+  PUSHBYTES46 = 46,
+  PUSHBYTES47 = 47,
+  PUSHBYTES48 = 48,
+  PUSHBYTES49 = 49,
+  PUSHBYTES50 = 50,
+  PUSHBYTES51 = 51,
+  PUSHBYTES52 = 52,
+  PUSHBYTES53 = 53,
+  PUSHBYTES54 = 54,
+  PUSHBYTES55 = 55,
+  PUSHBYTES56 = 56,
+  PUSHBYTES57 = 57,
+  PUSHBYTES58 = 58,
+  PUSHBYTES59 = 59,
+  PUSHBYTES60 = 60,
+  PUSHBYTES61 = 61,
+  PUSHBYTES62 = 62,
+  PUSHBYTES63 = 63,
+  PUSHBYTES64 = 64,
+  PUSHBYTES65 = 65,
+  PUSHBYTES66 = 66,
+  PUSHBYTES67 = 67,
+  PUSHBYTES68 = 68,
+  PUSHBYTES69 = 69,
+  PUSHBYTES70 = 70,
+  PUSHBYTES71 = 71,
+  PUSHBYTES72 = 72,
+  PUSHBYTES73 = 73,
+  PUSHBYTES74 = 74,
   PUSHBYTES75 = 75,
   PUSHDATA1 = 76,
   PUSHDATA2 = 77,
@@ -30,9 +103,8 @@ export enum OpCode {
   JMPIFNOT = 100,
   CALL = 101,
   RET = 102,
-  APPCALL = 103,
   SYSCALL = 104,
-  TAILCALL = 105,
+  DUPFROMALTSTACKBOTTOM = 105,
   DUPFROMALTSTACK = 106,
   TOALTSTACK = 107,
   FROMALTSTACK = 108,
@@ -84,13 +156,6 @@ export enum OpCode {
   MIN = 163,
   MAX = 164,
   WITHIN = 165,
-  SHA1 = 167,
-  SHA256 = 168,
-  HASH160 = 169,
-  HASH256 = 170,
-  CHECKSIG = 172,
-  VERIFY = 173,
-  CHECKMULTISIG = 174,
   ARRAYSIZE = 192,
   PACK = 193,
   UNPACK = 194,
@@ -105,11 +170,6 @@ export enum OpCode {
   HASKEY = 203,
   KEYS = 204,
   VALUES = 205,
-  CALL_I = 224,
-  CALL_E = 225,
-  CALL_ED = 226,
-  CALL_ET = 227,
-  CALL_EDT = 228,
   THROW = 240,
   THROWIFNOT = 241
 }
