+++ packages/neon-core/src/sc/index.ts
@@ -1,5 +1,5 @@
 export * from "./ContractParam";
 export * from "./core";
 export * from "./ScriptBuilder";
-export * from "./StackItem";
 export * from "./OpCode";
+export * from "./InteropService";
