--- packages/neon-core/src/sc_v3/index.ts
@@ -1,5 +1,0 @@
-export * from "./ContractParam";
-export * from "./core";
-export * from "./ScriptBuilder";
-export * from "./OpCode";
-export * from "./InteropService";
