+++ packages/neon-core/src/tx/components/index.ts
@@ -1,5 +1,2 @@
 export * from "./TransactionAttribute";
-export * from "./TransactionInput";
-export * from "./TransactionOutput";
 export * from "./Witness";
-export * from "./StateDescriptor";
