+++ packages/neon-core/src/tx/index.ts
@@ -1,5 +1,3 @@
 export * from "./components";
 export * from "./transaction";
-export * from "./strategy";
 export * from "./txAttrUsage";
-export { defaultCalculationStrategy } from "./calculate";
