+++ packages/neon-core/src/tx/transaction/index.ts
@@ -1,12 +1,1 @@
-export * from "./BaseTransaction";
 export * from "./Transaction";
-export * from "./ClaimTransaction";
-export * from "./ContractTransaction";
-export * from "./InvocationTransaction";
-export * from "./StateTransaction";
-export * from "./MinerTransaction";
-export * from "./IssueTransaction";
-export * from "./RegisterTransaction";
-export * from "./TransactionType";
-export * from "./EnrollmentTransaction";
-export * from "./PublishTransaction";
