--- packages/neon-core/src/tx_v3/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./TransactionAttribute";
-export * from "./Witness";
