--- packages/neon-core/src/tx_v3/index.ts
@@ -1,3 +1,0 @@
-export * from "./components";
-export * from "./transaction";
-export * from "./txAttrUsage";
