--- packages/neon-core/src/tx_v3/transaction/index.ts
@@ -1,1 +1,0 @@
-export * from "./Transaction";
