--- packages/neon-core/src/tx_v3/txAttrUsage.ts
@@ -1,10 +1,0 @@
-/**
- * Enum for TransactionAttributeUsage
- * @enum {number}
- */
-export enum TxAttrUsage {
-  Cosigner = 32,
-  Url = 129
-}
-
-export default TxAttrUsage;
