+++ packages/neon-core/src/wallet/Account.ts
@@ -3,10 +3,7 @@ import { DEFAULT_ACCOUNT_CONTRACT, DEFAULT_SCRYPT } from "../consts";
 import logger from "../logging";
 import { hash160, reverseHex } from "../u";
 import * as core from "./core";
-import {
-  constructMultiSigVerificationScript,
-  constructMultiSigVerificationScriptNeo3
-} from "./multisig";
+import { constructMultiSigVerificationScript } from "./multisig";
 import { decrypt, encrypt, ScryptParams } from "./nep2";
 import {
   isAddress,
@@ -61,31 +58,12 @@ export class Account {
    * ];
    * const acct = Account.createMultiSig(threshold, publicKeys);
    */
+
   public static createMultiSig(
     signingThreshold: number,
     publicKeys: string[]
   ): Account {
     const verificationScript = constructMultiSigVerificationScript(
-      signingThreshold,
-      publicKeys
-    );
-    return new Account({
-      contract: {
-        script: verificationScript,
-        parameters: Array(signingThreshold).map((_, i) => ({
-          name: `signature${i}`,
-          type: "Signature"
-        })),
-        deployed: false
-      }
-    });
-  }
-
-  public static createMultiSigNeo3(
-    signingThreshold: number,
-    publicKeys: string[]
-  ): Account {
-    const verificationScript = constructMultiSigVerificationScriptNeo3(
       signingThreshold,
       publicKeys
     );
